+++ src/lib.rs
@@ -13,7 +13,8 @@
 //!     let manager = Manager::new();
 //!
 //!     for loader in manager.get_loader_list().iter() {
-//!         let extensions: String = loader.get_extensions().join(",");
+//!         let extensions: Vec<String> = loader.get_extensions_mime_type_map().keys().map(|s| s.clone()).collect();
+//!         let extensions: String = extensions.join(",");
 //!         let mime_types: String = loader.get_mime_types().join(",");
 //!         println!(
 //!             "Loader {}: Extensions=[{}], Mime-Types=[{}] ",
@@ -45,7 +46,7 @@
 //!     } else {
 //!         let file_path = Path::new(&args[1]);
 //!
-//!         let cad_data = loader.read_file(&file_path);
+//!         let cad_data = loader.read_file(&file_path, mime_type);
 //!     }
 //! }
 //! ```