+++ src/loader/mod.rs
@@ -5,6 +5,6 @@ mod loader;
 mod manager;
 mod resource;
 
-pub use loader::Loader;
+pub use loader::{Loader, ExtensionMap};
 pub use manager::Manager;
 pub use resource::*;